refactor(tui): split settings_tests.rs into focused files under 300 lines each
settings_tests.rs (1173 lines) exceeded the 300-line file limit. Split into:
- helpers.rs: shared test utilities (render helpers, setup fns, file creators)
- settings_ui_tests.rs: keybindings, open/close, focus, layout render tests
- settings_export_tests.rs: export dialog state, handlers, rendering, execution
- settings_navigation_tests.rs: panel navigation (j/k/h/l/1/2/3), cli override skip
- settings_storage_tests.rs: file arg detection, storage switching, migration
- settings_config_edit_tests.rs: config editing, validation, prefix sync, migration state
